fix: WImage

+++ src/core/WImage.cpp
@@ -30,7 +30,9 @@ WImage::WImage(int width, int height, QImage::Format format) : QImage(width, hei
 }
 
 WImage::WImage(int page) :
-    QImage(Page::getResolutionWidth(), Page::getResolutionHeigth() * page, WImage::Format_ARGB32)
+    QImage(Page::getResolutionWidth(),
+           static_cast<int>(Page::getResolutionHeigth()) * page,
+           WImage::Format_ARGB32)
 {
     W_ASSERT(page > 0);
     const auto res = QImage::isNull();
